refactor(widgets): extract shared IconToggleButton
Three near-identical compact toggle buttons (ranking filter controls, railway
coverage sort controls, statistics sort button) are unified into a single
lib/widgets/icon_toggle_button.dart.

An `onCard` flag preserves the two looks exactly: card variant (opaque
surface/white fill, outlineVariant border, padding 9) for the ranking rows, and
scaffold variant (transparent fill, outline border, padding 8) for statistics.
`active` and `enabled` keep the highlight / disabled behaviours. Behaviour is
unchanged in every feature.
