+++ lib/features/ranking/widgets/coverage_sort_controls.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:trainlog_app/l10n/app_localizations.dart';
+import 'package:trainlog_app/widgets/icon_toggle_button.dart';
 
 /// The inline sorting toggles shared by the Railway Coverage list views and the
 /// drill-down page: an Alphabetical/Value toggle and an Ascending/Descending
@@ -31,7 +32,7 @@ class CoverageSortControls extends StatelessWidget {
     return Row(
       mainAxisSize: MainAxisSize.min,
       children: [
-        _IconToggle(
+        IconToggleButton(
           active: alphabetical,
           enabled: enabled,
           icon: alphabetical ? Icons.tag : Icons.sort_by_alpha,
@@ -41,7 +42,7 @@ class CoverageSortControls extends StatelessWidget {
           onTap: onToggleAlphabetical,
         ),
         const SizedBox(width: 8),
-        _IconToggle(
+        IconToggleButton(
           active: false,
           enabled: enabled,
           icon: descending ? Icons.arrow_upward : Icons.arrow_downward,
@@ -51,53 +52,6 @@ class CoverageSortControls extends StatelessWidget {
           onTap: onToggleDirection,
         ),
       ],
-    );
-  }
-}
-
-class _IconToggle extends StatelessWidget {
-  final bool active;
-  final bool enabled;
-  final IconData icon;
-  final String tooltip;
-  final VoidCallback onTap;
-
-  const _IconToggle({
-    required this.active,
-    required this.enabled,
-    required this.icon,
-    required this.tooltip,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final cs = Theme.of(context).colorScheme;
-    final isDark = Theme.of(context).brightness == Brightness.dark;
-    final cardColor = isDark ? cs.surface : Colors.white;
-
-    return Opacity(
-      opacity: enabled ? 1 : 0.4,
-      child: Tooltip(
-        message: tooltip,
-        child: GestureDetector(
-          onTap: enabled ? onTap : null,
-          child: Container(
-            padding: const EdgeInsets.all(9),
-            decoration: BoxDecoration(
-              color: active ? cs.primary.withValues(alpha: 0.15) : cardColor,
-              borderRadius: BorderRadius.circular(10),
-              border: Border.all(
-                color: active
-                    ? cs.primary.withValues(alpha: 0.6)
-                    : cs.outlineVariant.withValues(alpha: 0.4),
-                width: 1.2,
-              ),
-            ),
-            child: Icon(icon, size: 18, color: cs.primary),
-          ),
-        ),
-      ),
     );
   }
 }
